add presenter utility re #43
