+++ mslice/presenters/slice_plotter_presenter.py
@@ -1,10 +1,9 @@
 from __future__ import (absolute_import, division, print_function)
 from mslice.models.slice.slice_plotter import SlicePlotter
+from mslice.presenters.presenter_utility import PresenterUtility
 from mslice.views.slice_plotter_view import SlicePlotterView
 from mslice.widgets.slice.command import Command
-from mantid.api import IMDHistoWorkspace
 from .interfaces.slice_plotter_presenter import SlicePlotterPresenterInterface
-from .interfaces.main_presenter import MainPresenterInterface
 from .validation_decorators import require_main_presenter
 
 
@@ -34,7 +33,7 @@ INVALID_X_UNITS = 6
 INVALID_Y_UNITS = 7
 
 
-class SlicePlotterPresenter(SlicePlotterPresenterInterface):
+class SlicePlotterPresenter(PresenterUtility, SlicePlotterPresenterInterface):
     def __init__(self, slice_view, slice_plotter):
         if not isinstance(slice_view, SlicePlotterView):
             raise TypeError("Parameter slice_view is not of type SlicePlotterView")
@@ -46,14 +45,8 @@ class SlicePlotterPresenter(SlicePlotterPresenterInterface):
         colormaps = self._slice_plotter.get_available_colormaps()
         self._slice_view.populate_colormap_options(colormaps)
 
-
-    def register_master(self, main_presenter):
-        assert (isinstance(main_presenter, MainPresenterInterface))
-        self._main_presenter = main_presenter
-        self._main_presenter.subscribe_to_workspace_selection_monitor(self)
-
     def notify(self, command):
-        self._clear_displayed_error()
+        self._clear_displayed_error(self._slice_view)
         self._slice_view.busy.emit(True)
         if command == Command.DisplaySlice:
             self._display_slice()
@@ -146,16 +139,6 @@ class SlicePlotterPresenter(SlicePlotterPresenterInterface):
             if y.start > y.end:
                 return INVALID_Y_PARAMS
 
-    def _to_float(self, x):
-        if x == "":
-            return None
-        return float(x)
-
-    def _to_int(self, x):
-        if x == "":
-            return None
-        return int(x)
-
     def workspace_selection_changed(self):
         workspace_selection = self._get_main_presenter().get_selected_workspaces()
         if len(workspace_selection) != 1:
@@ -177,9 +160,6 @@ class SlicePlotterPresenter(SlicePlotterPresenterInterface):
             return
         self._slice_view.populate_slice_x_params(*["%.5f" % x for x in (x_min, x_max, x_step)])
         self._slice_view.populate_slice_y_params(*["%.5f" % x for x in (y_min, y_max, y_step)])
-
-    def _clear_displayed_error(self):
-        self._slice_view.clear_displayed_error()
 
     def set_workspace_provider(self, workspace_provider):
         self._slice_plotter.set_workspace_provider(workspace_provider)
